A Badminton Ladder using a Turtle GUI
A data management project with graphical display. The project is a badminton ladder management system which provide means for a player to take one of these actions:
a. Issue a challenge (stating opponent and play-by date)
b. Record the result of a challenge
c. Join the ladder
d. Withdraw from the ladder
e. Make a query
ii. Provide means of recording and displaying the result of a challenge.
iii. Adjust the ladder according to the result.
iv. Register a new player and place him/her on the ladder.
v. Accept the withdrawal of a player from the ladder and update the ladder accordingly.